Update AntiMicMonitoringDesync.py
Prior to disabling and re-enabling the monitoring source save the monitoring type so that it can be enabled using the users monitoring type that was previously set. Otherwise, it's forcing the monitoring type to be "Monitor" and "Output" and that may not always be how a user has the source configured.